google_chess_player_agent/pyproject.toml
[project]
name = "google-chess-player-agent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk>=0.3.2",
    "aiohttp>=3.12.15",
    "google-adk>=1.12.0",
    "python-chess>=1.999",
    "python-dotenv>=1.1.1",
    "urllib3>=2.5.0",
]
